pyproject.toml
[tool.poetry]
name = "gukbot-v2"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
python-dotenv = "^1.0.1"
slack-bolt = "^1.21.2"
slack-sdk = "^3.33.4"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
